+++ app/views/answers/_form.html.erb
@@ -1,0 +1,11 @@
+
+<%= form_for(@answer) do |f| %>
+    <%= f.hidden_field(:user_id, value: @answer.user_id) %>
+    <%= f.hidden_field(:question_id, value: @answer.question_id) %>
+    <%= f.hidden_field(:answer_type, value: @answer.answer_type) %>
+
+    <%= render(:partial => "question", :locals => {:question => @question, :f => f}) %>   
+
+    <%= link_to("Go Back", answers_path) %>
+    <%= f.submit("Save") %>
+<% end %>
