+++ app/views/answers/_numeric.html.erb
@@ -1,0 +1,31 @@
+<div class="d-flex justify-content-center">
+  <!-- 1 -->
+  <div class="form-check d-flex mx-4">
+    <%= f.radio_button(:numeric, "1", class: "form-check-input") %>
+    <%= f.label("1", class: "form-check-label") %>
+  </div>
+  
+  <!-- 2 -->
+  <div class="form-check d-flex mx-4">
+    <%= f.radio_button(:numeric, "2", class: "form-check-input") %>
+    <%= f.label("2", class: "form-check-label") %>
+  </div>
+  
+  <!-- 3 -->
+  <div class="form-check d-flex mx-4">
+    <%= f.radio_button(:numeric, "3", class: "form-check-input") %>
+    <%= f.label("3", class: "form-check-label") %>
+  </div>
+  
+  <!-- 4 -->
+  <div class="form-check d-flex mx-4">
+    <%= f.radio_button(:numeric, "4", class: "form-check-input") %>
+    <%= f.label("4", class: "form-check-label") %>
+  </div>
+  
+  <!-- 5 -->
+  <div class="form-check d-flex mx-4">
+    <%= f.radio_button(:numeric, "5", class: "form-check-input") %>
+    <%= f.label("5", class: "form-check-label") %>
+  </div>
+</div>
