+++ app/views/answers/_question.html.erb
@@ -1,0 +1,12 @@
+<!-- Card layout for each question plus CRUD for each. -->
+<div class="card mb-4">
+  <!-- Card Header -->
+  <div class="card-header d-flex justify-content-between">
+    <h5>Question <%= question.id %></h5>
+  </div>
+  <!-- Card Body -->
+  <div class="card-body">
+    <h5 class="card-title"><%= question.question %></h5>
+    <%= render(:partial => get_partial(question), :locals => {:question => question, :f => f}) %>
+  </div>
+</div>
