+++ app/views/answers/_short_answer.html.erb
@@ -1,0 +1,1 @@
+<%= f.text_area(:short_answer, class: "form-control", rows: "6") %>
