+++ app/views/answers/_true_false.html.erb
@@ -1,0 +1,12 @@
+<div>
+  <!-- True -->
+  <div class="form-check mx-4">
+    <%= f.radio_button(:true_false, true, class: "form-check-input") %>
+    <%= f.label("True", class: "form-check-label") %>
+  </div>
+  <!-- False -->
+  <div class="form-check mx-4">
+    <%= f.radio_button(:true_false, false, class: "form-check-input") %>
+    <%= f.label("False", class: "form-check-label") %>
+  </div>
+</div>
