+++ app/views/answers/index.html.erb
@@ -20,32 +20,15 @@
         <td class="index"><%= row_num %></td>
         <td><%= question.question %></td>
         <td>
-            <%= 
-                if @answers.exists?(:question_id => question.id)
-                    answer = @answers.where(:question_id => question.id).first
-
-                    case question.question_type
-                    when "Multiple Choice"
-                        "Unimplemented"
-                    when "True/False"
-                        "Unimplemented"
-                    when "Numeric"
-                        "Unimplemented"
-                    else
-                        answer.short_answer
-                    end
-                else
-                    "Unanswered"
-                end
-            %>
+            <%= get_answer_preview(question) %>
         </td>
         <td class="actions">
             <% if @answers.exists?(:question_id => question.id) %>
                 <% answer = @answers.where(:question_id => question.id).first %>
-                <%= link_to("Edit", edit_answer_path(answer, :question => question.question)) %>
-                <%= link_to("Delete Response", delete_answer_path(answer, :question => question.question)) %>
+                <%= link_to("Edit", edit_answer_path(answer, :question_id => question.id)) %>
+                <%= link_to("Delete Response", delete_answer_path(answer, :question_id => question.id)) %>
             <% else %>
-                <%= link_to("Add Response", new_answer_path(:question_id => question.id, :question => question.question, :answer_type => question.question_type)) %>
+                <%= link_to("Add Response", new_answer_path(:question_id => question.id)) %>
             <% end %>
         </td>
     </tr>
